shop owner role & admin role access set, header and foooter sections created as seperate entity, proper folder for css

+++ src/main/java/com/ppf/project/dao/DeviceDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ppf.project.model.Device;
 
-public interface DeviceDao extends JpaRepository<Device, Long> {
+public interface DeviceDao extends JpaRepository<Device, Integer> {
 
 }
